refactor(curric): moved components
