+++ src/components/Grid/CurriculumInfo/CurriculumHeader.tsx
