+++ src/components/Grid/CurriculumInfo/tabs/SequenceTab.tsx
